Move stable diffusion files into the source lib under new diffusers subfolder
