+++ notebooks/stable_diffusion/requirements.txt
@@ -1,3 +1,2 @@
-diffusers[torch]==0.11.1
 optimum-graphcore
 matplotlib
