+++ optimum/graphcore/diffusers/__init__.py
@@ -1,0 +1,7 @@
+from .pipelines import (
+    INFERENCE_ENGINES_TO_MODEL_NAMES,
+    IPUStableDiffusionImg2ImgPipeline,
+    IPUStableDiffusionInpaintPipeline,
+    IPUStableDiffusionPipeline,
+    get_default_ipu_configs,
+)
