+++ optimum/graphcore/diffusers/pipelines/__init__.py
@@ -1,0 +1,7 @@
+from .stable_diffusion import (
+    INFERENCE_ENGINES_TO_MODEL_NAMES,
+    IPUStableDiffusionImg2ImgPipeline,
+    IPUStableDiffusionInpaintPipeline,
+    IPUStableDiffusionPipeline,
+    get_default_ipu_configs,
+)
