+++ optimum/graphcore/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion.py
@@ -1,0 +1,7 @@
+from diffusers import StableDiffusionPipeline
+
+from .pipeline_stable_diffusion_mixin import IPUStableDiffusionPipelineMixin
+
+
+class IPUStableDiffusionPipeline(IPUStableDiffusionPipelineMixin, StableDiffusionPipeline):
+    pass
