+++ optimum/graphcore/diffusers/pipelines/stable_diffusion/pipeline_stable_diffusion_img2img.py
@@ -1,0 +1,7 @@
+from diffusers import StableDiffusionImg2ImgPipeline
+
+from .pipeline_stable_diffusion_mixin import IPUStableDiffusionPipelineMixin
+
+
+class IPUStableDiffusionImg2ImgPipeline(IPUStableDiffusionPipelineMixin, StableDiffusionImg2ImgPipeline):
+    pass
