+++ setup.py
@@ -15,6 +15,7 @@ except Exception as error:
 INSTALL_REQUIRES = [
     "transformers==4.25.1",
     "optimum==1.6.1",
+    "diffusers[torch]==0.11.1",
     "datasets",
     "tokenizers",
     "torch @ https://download.pytorch.org/whl/cpu/torch-1.13.1%2Bcpu-cp38-cp38-linux_x86_64.whl",
